requirements.txt
Django==3.2
djangorestframework
Pillow
python-decouple
psycopg2-binary

